requirements.txt
flask==3.1.1
flask-cors==6.0.1
pyjwt==2.10.1
requests==2.32.4
python-dotenv==1.0.1
pytest==7.4.0
pytest-cov==4.1.0
gunicorn==21.2.0
